% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{super_gather}
\alias{super_gather}
\title{Gather columns into key-values pairs.}
\usage{
super_gather(data, key = "key", key_names = NULL, value_names = NULL,
  name_order, delim_nchar = 1)
}
\arguments{
\item{data}{A data frame.}

\item{key}{Name of new key column, as string.}

\item{key_names}{Names of key levels present in column names to be gathered,
as character vector. Specify either this or \code{value_names} but not both.
If both this and \code{value_names} are specified, will \emph{only} use
key names to identify columns to gather.}

\item{value_names}{Names of value levels present in column names to be gathered,
as character vector. Specify either this or \code{key_names} but not both.}

\item{name_order}{Which identifier comes first in existing colnames? Used to assist in
identifying which columns to gather. Choose \code{"key_first"} or \code{"value_first"}.}

\item{delim_nchar}{How many characters delimit between key and value names? Defaults to 1.}
}
\value{
A data frame, with all indicated columns gathered, but separate.
}
\description{
Partially collapse multiple columns into key-values pairs. Use \code{super_gather()}
when you have have multiple wide-form variables that you want to partially collapse,
but you want those gathered variables to stay in separate columns. Equivalent to
calling \code{\link[tidyr]{gather}} separately for dataframes with the \emph{same key} and
different values, and joining them column-wise back into one dataframe. Requires the common
key(s) to be spelled identically in names of columns to be gathered.
}
\details{
Implements a strategy akin to calling \code{\link[tidyr]{gather}}, \code{\link[tidyr]{separate}}, and
\code{\link[tidyr]{spread}} in succession. As a result, may coerce some columns to character
under the hood. Please pay attention to warnings.
}
\examples{
data <- data.frame(id = 1:10,
value1_cond1 = "a",
value1_cond2 = "b",
value2_cond1 = 0L,
value2_cond2 = 1L)

super_gather(data, "condition", key_names = c("cond1", "cond2"), name_order = "value_first")
super_gather(data, "condition", value_names = c("value1", "value2"), name_order = "value_first")
}
