% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_math.R
\name{tidyselect_row_sums}
\alias{tidyselect_row_sums}
\alias{tidyselect_row_means}
\title{Calculate row sums & means tidyselect-safely.}
\usage{
tidyselect_row_sums(.data, ..., .value = "row_sum", na.rm = FALSE)

tidyselect_row_means(.data, ..., .value = "row_mean", na.rm = FALSE)
}
\arguments{
\item{.data}{A tbl: a \code{data.frame} or \code{tibble}.}

\item{...}{One or more unquoted variable names. Accepts \code{\link[tidyselect]{select_helpers}}.}

\item{.value}{The name of the new column, as a string.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
An object of the same class as \code{.data}, with the computed column added.
}
\description{
Wrapper around \code{\link[base]{rowSums}} etc. to allow
calculating row sums and means of flexibly supplied column names
from a df. Accepts \code{\link[tidyselect]{select_helpers}}.
Designed for external use in pipe-chains, \emph{outside} of \code{\link[dplyr]{mutate}}.
}
\examples{
data <- tibble::tibble(x1 = 1:5, x2 = 1:5, x3 = 1:5)

tidyselect_row_sums(data, dplyr::starts_with("x"))
tidyselect_row_means(data, dplyr::starts_with("x"))
}
