% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_math.R
\name{tidy_row_sums}
\alias{tidy_row_sums}
\alias{tidy_row_means}
\title{Calculate row sums & means tidyverse-safely.}
\usage{
tidy_row_sums(...)

tidy_row_means(...)
}
\arguments{
\item{...}{One or more unquoted variable names. Does not currently accept
\code{\link[tidyselect]{select_helpers}} (hopefully to be added soon).}
}
\value{
A vector of equal length to each of the input vectors/columns.
}
\description{
Wrapper around \code{\link[base]{rowSums}} etc. to allow
calculating row sums and means of flexibly supplied column names
from a df with \code{link[dplyr]{select}}-style syntax. Safe to
use inside of \code{\link[dplyr]{mutate}}.
}
